requirements.txt
notebook
ensure
mkdocs-material
ipywidgets
tqdm
python-dotenv
pip-chill
-e .
